VM: Widen poem top-level definition counts to `uint32`.
- This is in preparation for constants tables being unique per function, which would allow assembling a Lore program into a single `.poem` file.
- A schema/global variable/function count limit of `uint16` would be too small for truly large Lore programs, so the count has been widened to `uint32`.

+++ compiler/src/lore/compiler/poem/writer/PoemWriter.scala
@@ -23,14 +23,13 @@ object PoemWriter {
   private def writeEntries(poemFragment: PoemFragment)(implicit constantsTable: ConstantsTable): BytecodeWriter = {
     implicit val writer: BytecodeWriter = new BytecodeWriter()
 
-    // No schemas and global variables yet.
-    writer.writeUInt16(poemFragment.schemas.length)
+    writer.writeUInt32(poemFragment.schemas.length)
     poemFragment.schemas.foreach(PoemSchemaWriter.write)
 
-    writer.writeUInt16(poemFragment.globalVariables.length)
+    writer.writeUInt32(poemFragment.globalVariables.length)
     poemFragment.globalVariables.foreach(PoemGlobalVariableWriter.write)
 
-    writer.writeUInt16(poemFragment.functions.length)
+    writer.writeUInt32(poemFragment.functions.length)
     poemFragment.functions.foreach(PoemFunctionWriter.write)
 
     writer
